requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
alembic
redis
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-text-splitters
chromadb
sentence-transformers
PyPDF2
python-docx
pytest
pytest-asyncio
httpx
numpy
pydantic
pydantic-settings
groq
tiktoken
chardet
python-magic